requirements.txt
# Social Media Analytics Pipeline Dependencies
# Core data processing
pandas==2.1.4
numpy==1.24.3

# API clients
tweepy==4.14.0
facebook-sdk==3.1.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Database
psycopg2-binary==2.9.9

# ETL and scheduling
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1

# HTTP requests
requests==2.31.0
aiohttp==3.9.1

# Data validation and transformation
pydantic==2.5.0
marshmallow==3.20.1

# Logging and monitoring
python-dotenv==1.0.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
